+++ CHANGELOG.md
@@ -2,6 +2,16 @@ CHANGELOG
 =========
 
 
+4.7, 2021-03-15
+---------------
+- added command-line flag --no-terminal (issue #97);
+- improved detetcing existing messages during online sync;
+- improved querying Skype online service for more messages to sync (issue #93);
+- fixed data grid export ignoring current filter;
+- fixed parsing parts of HTML entities into emoticons;
+- fixed command-line create overwriting existing file, and not creating file if password not given.
+
+
 4.6, 2021-03-04
 ---------------
 - update existing contact information in database on live sync;